+++ HISTORY.rst
@@ -2,6 +2,13 @@
 History
 =======
 
+0.1.1 (2019-07-18)
+==================
+
+- Fixed a spurious RuntimeWarning.
+- Improved error message if the altitude or location is missing in an
+  input file.
+
 0.1.0 (2019-06-21)
 ==================